# services/gateway/src/lib/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "The Gateway API",
    "description": "The Gateway API for the Microservices",
    "version": "1.0.0"
  },
  "host": "localhost:50000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/restaurant/{id}": {
      "get": {
        "tags": [
          "Restaurant"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Restaurant"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "restaurant-name"
                },
                "description": {
                  "type": "string",
                  "example": "restaurant-desc"
                },
                "locationList": {
                  "type": "array",
                  "example": [
                    1.099,
                    49.443
                  ],
                  "items": {
                    "type": "number"
                  }
                },
                "address": {
                  "type": "string",
                  "example": "restaurant-address"
                },
                "openingHoursList": {
                  "type": "array",
                  "example": [
                    "12h-14h",
                    "19h-22h"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "phone": {
                  "type": "string",
                  "example": "restaurant-phone"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Restaurant"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/restaurant": {
      "get": {
        "tags": [
          "Restaurant"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Restaurant"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "restaurant-name"
                },
                "description": {
                  "type": "string",
                  "example": "restaurant-desc"
                },
                "locationList": {
                  "type": "array",
                  "example": [
                    1.099,
                    49.443
                  ],
                  "items": {
                    "type": "number"
                  }
                },
                "address": {
                  "type": "string",
                  "example": "restaurant-address"
                },
                "openingHoursList": {
                  "type": "array",
                  "example": [
                    "12h-14h",
                    "19h-22h"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "phone": {
                  "type": "string",
                  "example": "restaurant-phone"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/restaurant/by-location": {
      "post": {
        "tags": [
          "Restaurant"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "locationList": {
                  "type": "array",
                  "example": [
                    1.099,
                    49.443
                  ],
                  "items": {
                    "type": "number"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/promotion/{code}": {
      "get": {
        "tags": [
          "Promotion"
        ],
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/promotion": {
      "get": {
        "tags": [
          "Promotion"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Promotion"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "example": "code-example"
                },
                "reduction": {
                  "type": "string",
                  "example": "reduction-example"
                },
                "method": {
                  "$ref": "#/definitions/Method"
                },
                "restaurantId": {
                  "type": "string",
                  "example": "restaurant_id:0"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/promotion/by-restaurant/{restaurantId}": {
      "get": {
        "tags": [
          "Promotion"
        ],
        "description": "",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/promotion/{id}": {
      "put": {
        "tags": [
          "Promotion"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "example": "code-example"
                },
                "reduction": {
                  "type": "string",
                  "example": "reduction-example"
                },
                "method": {
                  "$ref": "#/definitions/Method"
                },
                "restaurantId": {
                  "type": "string",
                  "example": "restaurant_id:0"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Promotion"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/stock/outcomes/by-restaurant/{id}": {
      "post": {
        "tags": [
          "Stock"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "string",
                  "example": "2022-01-01T00:00:00.000Z"
                },
                "interval": {
                  "type": "string",
                  "example": "interval-example"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Method": {
      "type": "string",
      "enum": [
        "PERCENT",
        "VALUE"
      ]
    },
...
  }
}